Breakout PublicKeyCredentialRpEntity and make immutable
